WIP: writing overview of the AI Cloud
This is planned to describe both the AI Cloud pilot platform and the "new"
AI Cloud. This will be updated in the future as the two generations merge.
NB: This content is not yet included in the documentation webpage.
Remember to move the source of the illustration
docs/assets/img/ai-cloud-pilot-overview.odg somewhere outside the webpage
folder hierarchy when we have finalised it.